+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - fix ability to sync when type files don't reflect models
 - leverage Map instead of an object so don't need to prefix numeric keys with underscore to prevent them from being sorted by numeric value rather than by the order in which they were added to the map
 - renameTable migration helper
+- the key function passed to `sortBy` may now also return a DateTime, a CalendarDate, or a bigint
+- fix `sort` on bigint arrays
 
 ## 1.9.4